Removal of Summary and Available Nagios services button from Icinga2 UI
With respect to bug #705 the avail.cgi and summary.cgi are a part of
Nagios core services which are not found in Icinga2. Also was not able
to migrate the cgi from Nagios to Icinga as I was not able to find the
cgi.cfg file in Icinga setting folder (seen in Icinga repository
https://github.com/Icinga/icinga2/tree/master/etc/icinga) in the Icinga
installation in docker container. So commented out both the buttons from
UI by modfiying the graph_content.php file.